moved prettify
