+++ src/util/types.ts
@@ -1,3 +1,8 @@
 export type Result<T, E = Error> =
   | { ok: T; error: undefined }
   | { ok: undefined; error: E };
+
+export type Prettify<T> = {
+  [K in keyof T]: T[K];
+  // eslint-disable-next-line @typescript-eslint/ban-types
+} & {};
